requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
beautifulsoup4==4.11.2
bitarray==2.7.3
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.0
django-browser-reload==1.7.0
django-countries==7.5.1
django-filter==22.1
django-fontawesome-6==1.0.0.0
django-friendship==1.9.6
django-modelcluster==6.0
django-no-last-login==0.1.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-tailwind==3.5.0
django-treebeard==4.6.1
djangorestframework==3.14.0
dotenv==0.0.5
draftjs-exporter==2.1.7
drf-yasg==1.21.5
et-xmlfile==1.1.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
ethereum-abi-utils==0.4.7
ethereum-tester==0.1.0b3
ethereum-utils==0.6.2
frozenlist==1.3.3
hexbytes==0.3.0
html5lib==1.1
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
ipfshttpclient==0.7.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
l18n==2021.3
lru-dict==1.1.7
Markdown==3.3.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==1.0.0
netaddr==0.8.0
openpyxl==3.1.2
packaging==23.0
parsimonious==0.9.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
protobuf==4.22.1
psycopg2==2.9.5
pycryptodome==3.17
pylru==1.2.1
pyparsing==3.0.9
pyrsistent==0.19.3
pysha3==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
regex==2023.3.22
requests==2.28.2
rlp==3.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
telepath==0.3
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
varint==1.0.2
wagtail==4.2.1
web3==6.0.0
webencodings==0.5.1
websockets==10.4
Willow==1.4.1
yarl==1.8.2
zipp==3.15.0